pip_requirements.txt
absl-py==2.2.2
accelerate==1.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
apex==0.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
autocommand==2.2.2
backports.tarfile==1.2.0
beautifulsoup4==4.13.3
boto3==1.37.9
botocore==1.37.9
braceexpand==0.1.7
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cpm-kernels==1.0.11
cycler==0.12.1
dashscope==1.22.2
datasets==3.3.2
decorator==5.2.1
decord==0.6.0
deepspeed==0.16.6
diffusers==0.32.2
dill==0.3.8
distro==1.9.0
DistVAE==0.0.0b5
easydict==1.13
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.17.0
flash-attn==2.7.3
fonttools==4.58.4
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
gradio==5.20.1
gradio_client==1.7.2
groovy==0.1.2
grpcio==1.71.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
inflect==7.3.1
iniconfig==2.1.0
inquirerpy==0.3.4
ipython==8.34.0
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
jmespath==1.0.1
kiwisolver==1.4.8
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.3.0
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.3
numpy==1.26.0
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
openai==1.65.4
opencv-python-headless==4.11.0.86
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pfzy==0.3.4
pillow==10.4.0
pipdeptree==2.28.0
platformdirs==4.2.2
pluggy==1.5.0
proglog==0.1.12
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==4.21.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyrallis==0.3.1
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruff==0.9.10
s2wrapper @ file:///cpfs01/projects-HDD/cfff-01ff502a0784_HDD/public/yanghao/software/scaling_on_scales
s3transfer==0.11.4
safehttpx==0.1.6
safetensors==0.5.3
scikit-video==1.1.11
scipy==1.15.2
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.46.1
SwissArmyTransformer==0.4.12
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tokenizers==0.21.0
tomli==2.0.1
tomlkit==0.13.2
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.1
triton==3.2.0
typeguard==4.3.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
webdataset==0.2.111
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
xformers==0.0.29.post3
xfuser==0.4.3.post3
xxhash==3.5.0
yarl==1.18.3
yunchang==0.6.3.post1
zipp==3.21.0
